+++ lib/src/properties.dart
@@ -59,7 +59,7 @@ class DetailedGlobalToggleProperties<T> extends GlobalToggleProperties<T> {
   /// The final width of the space between the icons.
   ///
   /// May differ from the value passed to the switch.
-  final double dif;
+  final double spacing;
 
   /// The final size of the indicator.
   ///
@@ -69,10 +69,10 @@ class DetailedGlobalToggleProperties<T> extends GlobalToggleProperties<T> {
   /// The size of the switch exclusive the outer wrapper
   final Size switchSize;
 
-  Size get difSize => Size(dif, switchSize.height);
+  Size get spacingSize => Size(spacing, switchSize.height);
 
   const DetailedGlobalToggleProperties({
-    required this.dif,
+    required this.spacing,
     required this.indicatorSize,
     required this.switchSize,
     required super.position,