+++ lib/src/widgets/custom_animated_toggle_switch.dart
@@ -12,7 +12,7 @@ typedef CustomIndicatorBuilder<T> = Widget Function(
 typedef CustomWrapperBuilder<T> = Widget Function(
     BuildContext context, GlobalToggleProperties<T> global, Widget child);
 
-/// Custom builder for the dif section between the icons.
+/// Custom builder for the [spacing] section between the icons.
 typedef CustomSeparatorBuilder<T> = Widget Function(BuildContext context,
     SeparatorProperties<T> local, DetailedGlobalToggleProperties<T> global);
 
@@ -115,13 +115,13 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
   /// Callback for selecting a new value. The new [current] should be set here.
   final ChangeCallback<T>? onChanged;
 
-  /// Space between the "indicator rooms" of the adjacent icons.
-  final double dif;
+  /// Space between adjacent icons.
+  final double spacing;
 
   /// Builder for divider or other separators between the icons.
   /// Builder for divider or other separators between the icons.
   ///
-  /// The available width is specified by [dif].
+  /// The available width is specified by [spacing].
   ///
   /// This builder is supported by [IconArrangement.row] only.
   final CustomSeparatorBuilder<T>? separatorBuilder;
@@ -194,7 +194,7 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
     this.animationCurve = Curves.easeInOutCirc,
     this.indicatorSize = const Size(48.0, double.infinity),
     this.onChanged,
-    this.dif = 0.0,
+    this.spacing = 0.0,
     this.separatorBuilder,
     this.onTap,
     this.fittingMode = FittingMode.preventHorizontalOverlapping,
@@ -222,7 +222,7 @@ class CustomAnimatedToggleSwitch<T> extends StatefulWidget {
   })  : assert(foregroundIndicatorBuilder != null ||
             backgroundIndicatorBuilder != null),
         assert(separatorBuilder == null ||
-            (dif > 0 && iconArrangement == IconArrangement.row)),
+            (spacing > 0 && iconArrangement == IconArrangement.row)),
         super(key: key);
 
   @override
@@ -402,7 +402,7 @@ class _CustomAnimatedToggleSwitchState<T>
                 properties.switchSize.width -
                     properties.indicatorSize.width / 2) -
             properties.indicatorSize.width / 2) /
-        (properties.indicatorSize.width + properties.dif);
+        (properties.indicatorSize.width + properties.spacing);
     if (properties.textDirection == TextDirection.rtl) {
       result = widget.values.length - 1 - result;
     }
@@ -424,7 +424,7 @@ class _CustomAnimatedToggleSwitchState<T>
 
   @override
   Widget build(BuildContext context) {
-    double dif = widget.dif;
+    double spacing = widget.spacing;
     final textDirection = _textDirectionOf(context);
     final loadingValue = _animationInfo.loading ? 1.0 : 0.0;
 
@@ -477,29 +477,29 @@ class _CustomAnimatedToggleSwitchState<T>
                         assert(
                             constraints.maxWidth.isFinite ||
                                 (widget.indicatorSize.width.isFinite &&
-                                    dif.isFinite),
+                                    spacing.isFinite),
                             'With unbound width constraints '
-                            'the width of the indicator and the dif '
+                            'the width of the indicator and the spacing '
                             "can't be infinite");
                         assert(
-                            widget.indicatorSize.width.isFinite || dif.isFinite,
+                            widget.indicatorSize.width.isFinite || spacing.isFinite,
                             'The width of the indicator '
-                            'or the dif must be finite.');
+                            'or the spacing must be finite.');
 
                         // Recalculates the indicatorSize if its width or height is
                         // infinite.
                         Size indicatorSize = Size(
                             widget.indicatorSize.width.isInfinite
                                 ? (constraints.maxWidth -
-                                        dif * (widget.values.length - 1)) /
+                                        spacing * (widget.values.length - 1)) /
                                     widget.values.length
                                 : widget.indicatorSize.width,
                             widget.indicatorSize.height.isInfinite
                                 ? height
                                 : widget.indicatorSize.height);
 
-                        if (dif.isInfinite) {
-                          dif = (constraints.maxWidth -
+                        if (spacing.isInfinite) {
+                          spacing = (constraints.maxWidth -
                                   widget.indicatorSize.width *
                                       widget.values.length) /
                               (widget.values.length - 1);
@@ -508,7 +508,7 @@ class _CustomAnimatedToggleSwitchState<T>
                         // Calculates the required width of the widget.
                         double width =
                             indicatorSize.width * widget.values.length +
-                                (widget.values.length - 1) * dif;
+                                (widget.values.length - 1) * spacing;
 
                         // Handles the case that the required width of the widget
                         // cannot be used due to the given BoxConstraints.
@@ -516,7 +516,7 @@ class _CustomAnimatedToggleSwitchState<T>
                                 FittingMode.preventHorizontalOverlapping &&
                             width > constraints.maxWidth) {
                           double factor = constraints.maxWidth / width;
-                          dif *= factor;
+                          spacing *= factor;
                           width = constraints.maxWidth;
                           indicatorSize = Size(
                               indicatorSize.width.isInfinite
@@ -524,7 +524,7 @@ class _CustomAnimatedToggleSwitchState<T>
                                   : factor * indicatorSize.width,
                               indicatorSize.height);
                         } else if (constraints.minWidth > width) {
-                          dif += (constraints.minWidth - width) /
+                          spacing += (constraints.minWidth - width) /
                               (widget.values.length - 1);
                           width = constraints.minWidth;
                         }
@@ -538,7 +538,7 @@ class _CustomAnimatedToggleSwitchState<T>
 
                         // The local position of the indicator.
                         double position =
-                            (indicatorSize.width + dif) * positionValue +
+                            (indicatorSize.width + spacing) * positionValue +
                                 indicatorSize.width / 2;
 
                         double leftPosition = textDirection == TextDirection.rtl
@@ -562,7 +562,7 @@ class _CustomAnimatedToggleSwitchState<T>
 
                         DetailedGlobalToggleProperties<T> properties =
                             DetailedGlobalToggleProperties(
-                          dif: dif,
+                          spacing: spacing,
                           position: positionValue,
                           indicatorSize: indicatorSize,
                           switchSize: Size(width, height),
@@ -704,12 +704,12 @@ class _CustomAnimatedToggleSwitchState<T>
       BuildContext context, DetailedGlobalToggleProperties<T> properties) {
     return [
       ...Iterable.generate(widget.values.length, (i) {
-        double position = i * (properties.indicatorSize.width + properties.dif);
+        double position = i * (properties.indicatorSize.width + properties.spacing);
         return Positioned.directional(
           textDirection: _textDirectionOf(context),
-          start: i == 0 ? position : position - properties.dif,
+          start: i == 0 ? position : position - properties.spacing,
           width: (i == 0 || i == widget.values.length - 1 ? 1 : 2) *
-                  properties.dif +
+                  properties.spacing +
               properties.indicatorSize.width,
           height: properties.indicatorSize.height,
           child: widget.iconBuilder(
@@ -751,7 +751,7 @@ class _CustomAnimatedToggleSwitchState<T>
                 )),
             if (i < length - 1 && widget.separatorBuilder != null)
               SizedBox(
-                width: properties.dif,
+                width: properties.spacing,
                 child: Center(
                   child: widget.separatorBuilder!(
                       context, SeparatorProperties(index: i), properties),